Various changes and improvements
Including:
- Groups now can hold custom data, too.
  Currently used to store commands settings.
- Custom built LuaInterface.dll, supports more accurate error messages.
- Completly rewritten data classes, which now work in Lua as well.
- Server now has the member "Public", which stores if the server's
  listed on the official minecraft public server list and fList.
- Removed all objects but portals.
- Portals rewritten. One Portal now holds and handles both Regions.
- Portals get saved/loaded when saving the level.
- And lots of bugfixes, I guess ...

+++ World/Blocktype.cs
@@ -121,7 +121,7 @@ namespace obsidian.World {
 			return type;
 		}
 		
-		public static Blocktype FindById(int id) {
+		public static Blocktype FindById(byte id) {
 			if (id>Count) { return null; }
 			return blocktypes[id];
 		}
